% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datasets.R
\docType{data}
\name{Dataset}
\alias{Dataset}
\alias{autoimmunityDT}
\alias{SummaryDF_Disease_HLA}
\alias{SummaryDF_Disease_HLAString}
\alias{SummaryDF_Disease_Peptide}
\title{Compiled datasets of HLA class I and class II epitopes associated with autoimmune diseases.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 2806 rows and 24 columns.}
\usage{
autoimmunityDT

SummaryDF_Disease_HLA

SummaryDF_Disease_HLAString

SummaryDF_Disease_Peptide
}
\description{
Datasets were compiled from source files attached in the package.
}
\keyword{datasets}
