% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialPresentationAnalysis.R
\name{DifferentialPresentationAnalysis}
\alias{DifferentialPresentationAnalysis}
\alias{differentialPresentationAnalysis}
\title{Differential presentation analysis.}
\usage{
differentialPresentationAnalysis(fileNames, hlaClass, rankPerThr = 2,
  DPIThr = 0.5)
}
\arguments{
\item{fileNames}{The set of NetMHC result files to be parsed. See the document of \code{NetMHC}.}

\item{hlaClass}{A string, either "ClassI" or "ClassII".}

\item{rankPerThr}{A threshold for rank percentile. For HLA class I and II, 2 and 10 are recommended, respectively.}

\item{DPIThr}{A threshold of differential presentation index, or DPI. We recommend 0.5. In this case, peptides with DPIs higher than 0.5 and lower than -0.5 are considered "predisposing" and "protective" epitopes, respectively.}
}
\description{
NetMHC prediction results are parsed and the mean, maximum, and minimum rank percentiles among disease-predisposing and -protective HLA molecules are computed. Note: currently, only four-digit HLA alleles are used to determine HLA-disease association, and thus prediction results from two-digit HLAs are discarded.
}
