\name{dpa-package}
\Rdversion{1.1}
\alias{dpa-package}
\alias{dpa}
\docType{package}
\title{Dynamic Path Approach for Analyzing time-dependent simulation data}
\description{
A GUI or command-line operated data analysis tool, 
        for analyzing time-dependent simulation data in which
	multiple instantaneous or time-lagged relations are 
        assumed. This package uses Structural Equation Modeling 
        (the sem package). It is aimed to deal with time-dependent
	data and estimate whether a causal diagram fits data
        from an (agent-based) simulation model.
}
\details{
\tabular{ll}{
Package: \tab dpa\cr
Type: \tab Package\cr
Version: \tab 1.0-03\cr
Date: \tab 2010-10-15\cr
License: \tab LGPL Version 2 or later.\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Emile J.L. Chappin <E.J.L.Chappin@TuDelft.NL>
}
\references{
http://www.chappin.com
}
\examples{
#Start the GUI with:
dpa.start()
}