\name{dpa.analysis.options}
\Rdversion{1.1}
\alias{dpa.analysis.options}
\title{Select the data and interval for analysis}
\description{
         The function provides an option to perform analysis on the whole data at a time 
         or for every time step or for a specific time interval. The function also gives you 
         an option of performing analysis for specific time steps and not for all time steps in 
         the data.
}
\usage{
  dpa.analysis.options()
}
\references{
 http://www.chappin.com
}
\author{
 Emile J.L. Chappin
}
\examples{
#dpa.analysis.options()
}
