\name{dpa.data.authenticationCancel}
\Rdversion{1.1}
\alias{dpa.data.authenticationCancel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Cancel the authetication window}
\description{
     You can return to the load data page by clicking cancel button on authentication page 
     at any moment.
}
\usage{
dpa.data.authenticationCancel()
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}

