\name{dpa.data.checkData}
\Rdversion{1.1}
\alias{dpa.data.checkData}
\title{Check data}
\description{
The data loaded needs to be checked for the missing rows and missing data in the
existing rows as that may result to errors in analysis.Also the data need to be 
sorted before being put further for analysis.
}
\usage{
dpa.data.checkData()
}
\references{
 http://www.chappin.com
}
\author{
 Emile J.L. Chappin
}
\examples{
#dpa.data.checkData()
}
