\name{dpa.data.loadCancel}
\Rdversion{1.1}
\alias{dpa.data.loadCancel}
\title{ cancel loading of data}
\description{
    It will allow you to cancel the screen which gives an option to select 
    the data loading either from the disk or database. 
}
\usage{
dpa.data.loadCancel()
}
\references{
 http://www.chappin.com
}
\author{
 Emile J.L. Chappin
}