\name{dpa.data.loadData}
\Rdversion{1.1}
\alias{dpa.data.loadData}
\title{Load data}
\description{
   It generates a screen which asks you to select one of the two options 
   available i.e. data file loading either from the disk or from the database.
}
\usage{
dpa.data.loadData()
}
\references{
 http://www.chappin.com
}
\author{
 Emile J.L. Chappin
}
\examples{
#dpa.data.loadData()
}