\name{dpa.data.loadDataFromDatabase}
\Rdversion{1.1}
\alias{dpa.data.loadDataFromDatabase}
\title{Load data from database}
\description{
   The function will open up an authentication window to ask for the information from 
   an user which is required to be verified before connecting to a server. 
}
\usage{
dpa.data.loadDataFromDatabase()
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}
