\name{dpa.data.saveDataToDisk}
\Rdversion{1.1}
\alias{dpa.data.saveDataToDisk}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Save data to disk}
\description{
    This function allows you to save the data loaded in the workspace to a   
    destination folder on disk selected through a folder browsing window. 
}
\usage{
dpa.data.saveDataToDisk(dataFileName=NULL)
}
\arguments{
  \item{dataFileName}{filename to which the data is saved}
}
\references{
 http://www.chappin.com
}
\author{
 Emile J.L. Chappin
}
\examples{
#dpa.saveDataToDisk()
}
