\name{dpa.data.setWorkingDirectory}
\Rdversion{1.1}
\alias{dpa.data.setWorkingDirectory}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Set working directory}
\description{
It pops up a folder browsing window in the GUi through which an user can select a particular directory to work with.
}
\usage{
dpa.data.setWorkingDirectory()
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}
\examples{
#dpa.data.setWorkingDirectory()
}