\name{dpa.data.viewOrEditData}
\Rdversion{1.1}
\alias{dpa.data.viewOrEditData}
\title{View or Edit data}
\description{
    This function allows to edit the loaded file in R editor in order to just have a look
    at the data or to make some changes in the data. Please make a note that the change is
    applied only to the loaded data and not to the data saved in the disk or database.	
}
\usage{
dpa.data.viewOrEditData()
}
\references{
 http://www.chappin.com
}
\author{
 Emile J.L. Chappin
}
\examples{
#dpa.data.viewOrEditData()
}