\name{dpa.incrementValue}
\Rdversion{1.1}
\alias{dpa.incrementValue}
\title{DPA increment value}
\description{
Increments the global value i
}
\usage{
dpa.incrementValue(i)
}
\arguments{
  \item{i}{The global value that is incremented}
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}
\examples{
#dpa.incrementValue(i)
}