\name{dpa.locate.missing}
\Rdversion{1.1}
\alias{dpa.locate.missing}
\title{Locate missing data}
\description{Locates and reports on missing data}
\usage{dpa.locate.missing(dataframe = NULL, tickColumn = NULL, jobColumn = NULL)}
\arguments{
  \item{dataframe}{Dataset}
  \item{tickColumn}{Column that represents time}
  \item{jobColumn}{Column that represents run or job}
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}
\examples{
#dpa.locate.missing(dataframe = NULL, tickColumn = NULL, jobColumn = NULL)
}