% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPA.R
\docType{package}
\name{DPA}
\alias{DPA}
\alias{package-DPA}
\alias{DPA-package}
\title{DPA: Differential presentation analysis for prioritizing disease-associated T cell epitopes}
\description{
DPA: Differential presentation analysis for prioritizing disease-associated T cell epitopes
}
