\name{dpa.relations.saveRelations}
\Rdversion{1.1}
\alias{dpa.relations.saveRelations}
\title{Save the relation}
\description{
    This function will allow the user to save the relation loaded in the 
    workspace to a destination folder on disk selected through a browsing 
    folder option.
}
\usage{
dpa.relations.saveRelations(saveRelFileName=NULL)
}
\arguments{
\item{saveRelFileName}{File to which the relations are saved}
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}
\examples{
#dpa.relations.saveRelations(saveRelFileName=NULL)
}