\name{dpa.results.setGraphDir}
\Rdversion{1.1}
\alias{dpa.results.setGraphDir}
\title{Change results directory}
\description{
The user is given an option to change the directory other than the working directory
for saving the results.
}
\usage{
dpa.results.setGraphDir(graphDir=NULL)
}
\arguments{
\item{graphDir}{Folder to set}
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}
\examples{
#dpa.results.setGraphDir(graphDir=NULL)
}