\name{dpa.results.viewNodePlots}
\Rdversion{1.1}
\alias{dpa.results.viewNodePlots}
\title{Generate graphs in between the parameters of data}
\description{
      This function designs a screen asking an user for the x and y parameters between 
      which the graph is to be plotted. Also there is an option to plot the graph for 
      all the time steps or for a specific time step as entered. The graph is generated 
      on clicking plot button and saved in the results directory.   
}
\usage{
dpa.results.viewNodePlots()
}
\details{
     It uses scatter plot function for generating the graphs. The graph is saved as
     .png file.  
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}