\name{dpa.results.viewRelationsPlots}
\Rdversion{1.1}
\alias{dpa.results.viewRelationsPlots}
\title{Generate the Relations plot }
\description{
        This function generates the plot of the relations showing every connection between 
        parameters along with their strength. The stronger a relation is the wider will be 
        a line connecting them. The parameter estimates value as got from the analysis result
        is also written at every connecting line after approximating it to 3 places after decimel.
  
}
\usage{
dpa.results.viewRelationsPlots(tickNumber=NULL)
}
\arguments{
\item{tickNumber}{if the analysis is performed per time step, that is denoted in the graph by specifying this parameter}
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}