\name{dpa.sort.data}
\Rdversion{1.1}
\alias{dpa.sort.data}
\title{Sort data}
\description{Sorts the dataset according to run or job first and time second.}
\usage{
dpa.sort.data(dataframe = NULL, tickColumn = NULL, runColumn = NULL)
}
\arguments{
  \item{dataframe}{The dataset to be sorted}
  \item{tickColumn}{The column depicting time}
  \item{runColumn}{The column depicting job or run}
}
\references{
http://www.chappin.com
}
\author{
Emile J.L. Chappin
}
\examples{
#dpa.sort.data(dataframe = NULL, tickColumn = NULL, runColumn = NULL)
}