% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetMHC.R
\name{NetMHC}
\alias{NetMHC}
\alias{NetMHCScript_HLAI}
\alias{NetMHCScript_HLAII}
\alias{NetMHC_Import_HLAI}
\alias{NetMHC_Import_HLAII}
\title{NetMHC analysis utility.}
\usage{
NetMHCScript_HLAI(dt, hlaDigit = "Four", outDir = "./NetMHC/",
  outputFileName = "NetMHCScript_HLA-I.txt")

NetMHCScript_HLAII(dt, hlaDigit = "Four", outDir = "./NetMHC/",
  outputFileName = "NetMHCScript_HLA-II.txt")

NetMHC_Import_HLAI(fileName)

NetMHC_Import_HLAII(fileName)
}
\arguments{
\item{dt}{A data.table of peptides, associated diseases and HLA strings. Note: NetMHCpan and NetMHCIIpan only accept peptides of 8~14aa and 9~32aa, respectively. Our script-generating functions remove peptides of other lengths intetrnally.}

\item{hlaDigit}{The digit of HLA alleles from PheWAS to be included in the NetMHC prediction. Either "Two", "Four", or "TwoAndFour". Note: all possible matched four-digit HLA alleles are used in place of two-digits HLA alleles.}

\item{outDir}{A path to the output directory for NetMHC analysis.}

\item{outputFileName}{A file name for the generated NetMHC script.}

\item{fileName}{A NetMHC result file to be parsed. It should be named like <Class>_<Disease>_<Length>. Eg. HLAI_AS_9mer.xls}
}
\description{
NetMHC analysis utility.
}
