% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Selfness.R
\name{Selfness}
\alias{Selfness}
\alias{Selfness_Features_HLAI}
\alias{Selfness_Features_HLAII}
\alias{Selfness_Score}
\alias{Selfness_Score_Extrapolation}
\title{Selfness score.}
\usage{
Selfness_Features_HLAI(peptideSet)

Selfness_Features_HLAII(peptideSet)

Selfness_Score(featureDT, sampling = NULL, seedSet = 1:5, mtry = 16,
  numRandomCuts = 16)

Selfness_Score_Extrapolation(featureDT, modelList, seedSet = 1:5)
}
\arguments{
\item{peptideSet}{A set of peptide sequences.}

\item{featureDT}{A peptide feature data.table.}

\item{sampling}{A sampling method passed to caret::trainControl.}

\item{seedSet}{A set of random seeds for ERT computation.}

\item{mtry}{A parameter passed to extraTrees::extraTrees.}

\item{numRandomCuts}{A parameter passed to extraTrees::extraTrees.}

\item{modelList}{A list of trained ERT models.}
}
\description{
Selfness score.
}
