% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniProtAlignment.R
\name{UniProtAlignment}
\alias{UniProtAlignment}
\alias{UniProt_Proteome_Import}
\alias{UniProt_Proteome_ExactMatch}
\alias{UniProt_Proteome_AlignmentScores}
\title{Sequence alignment-based homology scoring to UniProt proteomes.}
\usage{
UniProt_Proteome_Import(fastaFileName)

UniProt_Proteome_ExactMatch(peptideSet, proteome,
  coreN = parallel::detectCores(logical = F))

UniProt_Proteome_AlignmentScores(peptideSet, proteome,
  coreN = parallel::detectCores(logical = F))
}
\arguments{
\item{fastaFileName}{A UniProt proteome fasta file name to be parsed.}

\item{peptideSet}{A set of peptides.}

\item{proteome}{A set of protein sequences parsed by \code{UniProt_Proteome_Import}.}

\item{coreN}{The number of cores to be used for parallelization.}
}
\description{
Sequence alignment-based homology scoring to UniProt proteomes.
}
